% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Find a light threshold using calibration position and date}
\usage{
calibrate(df, day, lon, lat, zenith = 96, offset = 0, verbose = T)
}
\arguments{
\item{df}{data.frame containing `Light` and `Date` data at minimum}

\item{day}{POSIXct date-time object in GMT or UTC specifying one day of calibration data}

\item{lon}{known longitude of tag during calibration}

\item{lat}{known latitude of tag during calibration}

\item{zenith}{assumed solar zenith angle in degrees at twilight (depends on latitude and time of year, defaults to 96 degrees)}

\item{offset}{time in hours to offset start of day time, defaults to 0}

\item{verbose}{will plot light traces if TRUE}
}
\value{
threshold value for TwilightFree model
}
\description{
Find a light threshold using calibration position and date
}
