% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTracks.R
\name{drawTracks}
\alias{drawTracks}
\title{Plot tracks with map}
\usage{
drawTracks(gls, gps = NULL, point.cols = TRUE, line.col = "grey60",
  main = "", pacific = F)
}
\arguments{
\item{gls}{a trip object returned from `trip()` with `Lon` and `Lat` columns in time order at minimum}

\item{gps}{optional data.frame of reference data with `Lon` and `Lat` columns in time order}

\item{point.cols}{optional gradient colours for points, NULL if points not required}

\item{main}{optional title}

\item{pacific}{TRUE if map uses Pacific-centred co-ordinates}

\item{col}{line colour for trip}

\item{gps.col}{line colour for gps}
}
\description{
Plot tracks with map
}
