% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eraseLight.R
\name{eraseLight}
\alias{eraseLight}
\title{Manually erase non-solar light sources using mouse}
\usage{
eraseLight(obs, offset = 0, threshold = 5)
}
\arguments{
\item{obs}{the data frame containing at minimum Date and Light data}

\item{offset}{an offset in hours used to centre night within the plot y-axis}

\item{threshold}{ambient light threshold at which twilight is assumed to occur}

\item{zlim}{the dynamic range of the plot}
}
\value{
data frame with edited Light data
}
\description{
Click on the bottom-left and top-right corners of a box bounding the light you wish to erase
This function is slow, please be patient.
}
