% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findColony.R
\name{findColony}
\alias{findColony}
\title{Find days where animal was likely to have returned to colony}
\usage{
findColony(obs, threshold, colony, q = 0.9)
}
\arguments{
\item{obs}{a data frame containing at minimum Date and Light data}

\item{threshold}{numeric threshold for Light at twilights}

\item{colony}{vector with Lon and Lat coordinates for colony}

\item{q}{proportion of Light observations below threshold (between [0, 1])}
}
\value{
data frame of dates and location to use in `fixd` argument of `TwilightFree`
}
\description{
Where Light observations for a day are mostly below threshold (dark)
the animal is likely to have returned to a colony, nest, or burrow.
}
