% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMissing.R
\name{findMissing}
\alias{findMissing}
\title{Find days where no locations could be estimated from fitted TwilightFree object}
\usage{
findMissing(fit, type = "full")
}
\arguments{
\item{fit}{a fitted model returned by SGAT::Essie}

\item{type}{forward, backward or full posterior (defaults to "full")}
}
\value{
vector of Day_num
}
\description{
Find days where no locations could be estimated from fitted TwilightFree object
}
