% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findZenith.R
\name{findZenith}
\alias{findZenith}
\title{Find zenith angles from assumed threshold and locations}
\usage{
findZenith(d, day, lon, lat, offset = 0, threshold = 5)
}
\arguments{
\item{d}{data frame with Light, Date columns at minimum}
}
\value{
numeric zenith angle
}
\description{
Find zenith angles from assumed threshold and locations
}
