% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanLight.R
\name{interpolate_max_light}
\alias{interpolate_max_light}
\title{Find max light observations over a coarse time grid then interpolate at finer grid}
\usage{
interpolate_max_light(d, depth1 = FALSE, depth2 = FALSE,
  period1 = "30 minutes", period2 = "4 minutes")
}
\value{
see help(max_light_delta) for relevant details
}
\description{
Find max light observations over a coarse time grid then interpolate at finer grid
}
