% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{latentClass}
\alias{latentClass}
\title{Fit latent class mixture model to quickly classify day and night states}
\usage{
latentClass(d, nstates = 3L, verbose = TRUE)
}
\arguments{
\item{d}{data frame with Light, Date, delta columns at minimum}
}
\value{
data frame with predicted day and night states in `state` column
}
\description{
Fit latent class mixture model to quickly classify day and night states
}
