% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwilightFree.R
\name{logp0}
\alias{logp0}
\title{calculate SST component of log-posterior in TwilightFree model}
\usage{
logp0(k, x, slices)
}
\description{
calculate SST component of log-posterior in TwilightFree model
}
