% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrid.R
\name{makeGrid}
\alias{makeGrid}
\title{Make a grid for TwilightFree model incorporating land or sea mask}
\usage{
makeGrid(lon = c(-180, 180), lat = c(-90, 90), cell.size = 1,
  mask = "sea", pacific = FALSE)
}
\arguments{
\item{lon}{extents of longitude}

\item{lat}{extents of latitude}

\item{cell.size}{size of grid cells in degrees, defaults to 1}

\item{mask}{set mask for areas animals are assumed to be restricted to, "sea", "land", or "none"}

\item{pacific}{optional Pacific-centred map, defaults to Atlantic-centred map}
}
\value{
raster object giving the locations the animal may have visited
}
\description{
Make a grid for TwilightFree model incorporating land or sea mask
}
