% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPlot.R
\name{missingPlot}
\alias{missingPlot}
\title{Remove days from data where no locations could be estimated from fitted TwilightFree object}
\usage{
missingPlot(x, fit = NULL, threshold = 5, fixd = NULL, offset = 0)
}
\arguments{
\item{x}{data frame with Light, Date, and Day_num columns at minimum}

\item{fit}{a fitted model returned by SGAT::Essie}

\item{threshold}{threshold for twilight}

\item{fixd}{data frame with Date, Lon, Lat of fixed locations}
}
\value{
ts plot
}
\description{
Remove days from data where no locations could be estimated from fitted TwilightFree object
}
