% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberDays.R
\name{numberDays}
\alias{numberDays}
\title{Number each day in a column named Day_num}
\usage{
numberDays(obs)
}
\arguments{
\item{obs}{a data frame containing at minimum Date and Light data}
}
\value{
data frame with Day_num column of integers appended
}
\description{
A small number of TwilightFree helper functions require a `Day_num` column
}
