% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMissing.R
\name{removeMissing}
\alias{removeMissing}
\title{Remove days from data where no locations could be estimated from fitted TwilightFree object}
\usage{
removeMissing(x, fit)
}
\arguments{
\item{x}{data frame with Light, Date, and Day_num columns at minimum}

\item{fit}{a fitted model returned by SGAT::Essie}
}
\value{
data frame with inestimable days removed
}
\description{
Remove days from data where no locations could be estimated from fitted TwilightFree object
}
