% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdPlot.R
\name{thresholdPlot}
\alias{thresholdPlot}
\title{Plot light record using threshold to identify above-threshold light where we might reasonably expect dark}
\usage{
thresholdPlot(x, threshold = 5, offset = 0)
}
\arguments{
\item{x}{data frame with, at minimum, columns names `Date` and `Light`}

\item{threshold}{the pre-determined threshold of luminance which separates day from night}

\item{offset}{an offset in hours used to centre night within the plot y-axis}
}
\value{
plot of thresholded light
}
\description{
Plot light record using threshold to identify above-threshold light where we might reasonably expect dark
}
