% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{trip}
\alias{trip}
\title{Convert SGAT::essie fit object to matrix of `Date`, `Lon`, `Lat` columns}
\usage{
trip(fit, type = c("full", "forward", "backward"))
}
\arguments{
\item{fit}{SGAT::essie fit object}

\item{type}{"full" posterior, "forward" or "backward" probabilities to diagnose convergence of fit}
}
\value{
matrix with columns named `Date`, `Lon`, `Lat`
}
\description{
Convert SGAT::essie fit object to matrix of `Date`, `Lon`, `Lat` columns
}
