% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwilightFree.R
\name{TwilightFree}
\alias{TwilightFree}
\title{Specify a model for forwards-backwards estimation}
\usage{
TwilightFree(df, alpha = c(1, 1/10), beta = c(1, 1/4), dt = NULL,
  threshold = 5, zenith = 96, deployed.at = F, retrieved.at = F,
  fixd = NULL, sst = NULL)
}
\arguments{
\item{df}{data.frame containing `Light`, `Date`, and optionally `Temp` data}

\item{alpha}{hyperparameters for the noise (shading) assumption}

\item{beta}{hyperparameters for the movement assumption}

\item{dt}{optional parameter specifying the number of seconds in a segment (day)}

\item{threshold}{tag-specific value for luminance at twilight (obtained by calibration)}

\item{zenith}{solar zenith angle at twilight}

\item{deployed.at}{deployment location c(lon, lat) for first day of observation}

\item{retrieved.at}{retrieval location c(lon, lat) for last day of observation}

\item{fixd}{optional data.frame of fixed (known) locations containing `Date`, `Lon`, `Lat` (will overwrite deployed.at and retrieved.at locations if != NULL)}

\item{sst}{raster of SST data from NOAA OI SST}
}
\value{
a TwilightFree model object which can be fitted using SGAT::essie()
}
\description{
Specify a model for forwards-backwards estimation
}
