% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTracks.R
\name{drawTracks}
\alias{drawTracks}
\title{Plot tracks with map}
\usage{
drawTracks(trip, gps = NULL, col = "firebrick", gps.col = "dodgerblue",
  point.cols = c("red", "springgreen", "royalblue"), main = "",
  pacific = F)
}
\arguments{
\item{trip}{a trip object returned from `trip()`}

\item{gps}{optional data.frame of reference data with `Lon` and `Lat` columns in time order}

\item{col}{line colour for trip}

\item{gps.col}{line colour for gps}

\item{point.cols}{optional gradient colours for points, NULL if points not required}

\item{main}{optional title}

\item{pacific}{TRUE if map uses Pacific-centred co-ordinates}
}
\description{
Plot tracks with map
}
