% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eraseMoon.R
\name{eraseMoon}
\alias{eraseMoon}
\title{Remove days from data where no locations could be estimated from fitted TwilightFree object}
\usage{
eraseMoon(d, return.model = FALSE, model = NULL, q = 0.25)
}
\arguments{
\item{d}{data frame with Light, Date, Day_num, and state columns at minimum}

\item{return.model}{logical indicating whether to return predictive model or not}

\item{model}{a previously fitted predictive model}

\item{q}{number between 0 and 1 which controls sensitivity}
}
\value{
data frame with predicted night and day states in `state` column
}
\description{
Remove days from data where no locations could be estimated from fitted TwilightFree object
}
